Basic code for performing filtration of lists of tiddlers.
A filter string (in the format used in the <<list>> macro)
is translated into a list of functions which filter the
entire collection of tiddlers provided to the intial 
request. The resulting list of found tiddlers is created
in an additive fashion. It's not actually clear from the 
filter spec (the notional one) if that's the desired thing.

The upshot of this code is that higher level code can say:

GET me all the tiddlers in this bag that match the following
filter.

Which is what a recipe line is all about, so we're on our
way.

git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@3572 bb0f57cd-c710-0410-a8fb-e8bc1be0d679