Adjust filter syntax and code to support the concept of 'remove'. Keeping around the negate syntax as well.
So now we can do things like this, reading left to right:

  [!tag[mumbly]] [-tag[tagone]] TiddlerOne

Start a list of tiddlers with all those tiddlers in the 
base set which are not tagged with 'mumbly'. From that
resulting list remove all those tiddlers tagged with
'tagone' and add 'TiddlerOne'.

This is done by composing functions, so should be 
extensible. It's sort of a little filter string
compiler.

git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@3629 bb0f57cd-c710-0410-a8fb-e8bc1be0d679