very simple, stubbish cooking of a working tiddlywiki from a bag+filter style of recipe.
A recipe can now be serialized as a wiki, meaning its
bags are processed and filtered for the tiddlers within.
These tiddlers are then squirted in the right place in
empty.html.

test/make_a_wiki.py st?imulates this process and then 
outputs the tiddlywiki file to stdout, where it can be
saved and then opened in a browser et voila, there's the
tiddlers from the recipe.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@3831 bb0f57cd-c710-0410-a8fb-e8bc1be0d679