Decouple saving a tiddler from saving a bag.
We want this because only special people should be able to create
a bag and manipulate its security policy. However plenty of people
should be able to save to a bag that already exists.

So a tiddler needs to know what bag it belongs to. When a tiddler
is added to a bag it has its bag field automatically updated with
the name of the bag of which it is a member.

This will probably have some cascading effects which we'll sort
out later.



git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@3854 bb0f57cd-c710-0410-a8fb-e8bc1be0d679