Adjustments from playing with a large(r) datastore.
Move recipe.get_tiddlers() to control.get_tiddlers_from_recipe()
as we need to be able to pass in a store to the method to say
"really, get this stuff out of the store, not just thin air".

There will probably be a few more changes like this. The 
serializers are likely going to get complicated now. Such
is the way of persistence.

test/test_muchdata.py is a playground to learn from storing
and reading data (and it is working nicely c.f. above).

Next: See what it takes to make a good wiki from disk in
test_muchdata.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@3878 bb0f57cd-c710-0410-a8fb-e8bc1be0d679