tiddlyweb - add support for store and presenting revisions of tiddlers through tiddlyweb.
Nothing is done as yet on the client side to support handling revisions,
and nothing is done on the server to support contention mgt. That's to
come.

The fundamental change here is to the text storage system. Now instead
of a flat textfilefor storing a tiddler, we have a directory. In that
directory are files with numbers for the name. The number is the revision.
It starts at 1 and goes up.

The store API has only one change: list_tiddler_revisions(tiddler) lists the 
revisions of the provided tiddler. Access to a specific revision of a tiddler
is done by creating a new Tiddler object, setting the revision attribute and
doing store.get(tiddler). The store is expected to look at the revision field
and respect it if it is there. Otherwise, we desire the most recent revision.
This was deemed to be the most straightforward way to maintain a clean store
API that can easily be duplicated into other storage systems.

PUTting a tiddler always involves creating a new revision.

There are four new URLs on which a GET may be performed:

/recipes/{recipe_name}/tiddlers/{tiddler_name}/revisions
/recipes/{recipe_name}/tiddlers/{tiddler_name}/revisions/{revision_id}
/bags/{bag_name}/tiddlers/{tiddler_name}/revisions
/bags/{bag_name}/tiddlers/{tiddler_name}/revisions/{revision_id}

Next: 

* Experiment with TiddlyWebAdaptorPlugin to support revisions
  and collisions. This will likely require some server side
  changes in the HTTP response realm.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@4997 bb0f57cd-c710-0410-a8fb-e8bc1be0d679