tiddlyweb - adjust the store and serialization code so that PolicyS are properly saved and retrieved.
In a bit of a departure from previous behavior, policies are
written to disk in the text store as JSON dictionaries. This is
what could be called a "developer optimization", i.e. I want to
get on with using these policies not farting around with parsing
them etc. I figure since we're using JSON elsewhere as a requirement,
its groovy.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5172 bb0f57cd-c710-0410-a8fb-e8bc1be0d679