tiddlyweb - clean up exception handling a bit so that we:
* write to stderr when there is an uncaught exception
* catch exceptions at the highest point in the wsgi middleware stack

Before this change tracebacks and the like were being reported 
to the browser, but not to any kind of server logging. This is
a bit frustrating at times. So now fixed.

To make sure this is available to the entire system, teh HTTPExceptor
middleware is now called first, that is, it wraps the entire WSGI
application.



git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5404 bb0f57cd-c710-0410-a8fb-e8bc1be0d679