tiddlyweb - clean up the importer and associated code so that it easier to import the tiddlers from a tiddlywiki.html file.
The importer needs to make sure to decode &amp; and the like before
it sends data through a PUT to the server.

The JSON serialization must not set the attributes of tiddlers
to null or None but instead skip them. Without this, if the modification
key in a tiddler in JSON was present but had no value, then
the resulting tiddler on disk would have no modified value.

When we store a bag, be sure to set its policy owner so the policy
is properly formed.

At some point we can probably wire the importer into a web service
that sits alongside or within TiddlyWeb.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5405 bb0f57cd-c710-0410-a8fb-e8bc1be0d679