tiddlyweb - adjust the HTML serialization so that the generated HTML is more browseable. This gets us pretty close to a read only wiki for recipes and bags. There are even links that say "Recent Changes" now (when a single tiddler is being viewed in html you also get a link to the recent changes in the recipe or bag, depending on context).
When viewing a a list of tiddlers a "these tiddlers
as a tiddlywiki" link is provided.

Depending on the the setup of the tiddlyweb, the
generated tiddlywiki may or may not be edited. This
raises the question: should all TiddlyWiki's generated
by TiddlyWeb include the TiddlyWebAdaptorPlugin?
The TiddlyWebAutoSavePlugin?

Should there be a server side edit link/editor?

I'll update http://peermore.com:8080/ so people can
play with this.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5425 bb0f57cd-c710-0410-a8fb-e8bc1be0d679