tiddlyweb - add functionality and some tests so that it is possible to load serializations and stores from outside the tiddlyweb code repo. There's nothing around to actually hook this up, but this in preparation for that.
The ideas is that someone should be able to declare that say, the atom
type is available in module foo.bar.atom and support things properly.

This will take some doing to get the configuration handling going, but
this part is a step in the right direction.

Similar handling is going to be available for user auth.

git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5542 bb0f57cd-c710-0410-a8fb-e8bc1be0d679