tiddlyweb - Changes related to doing credentials extraction and authentication.
In order to move testing forward, we need to be able to
pass somewhat faked authentication information in requests.
So now the somewhat fake Credentials Extractor in
tiddlyweb.web.serve looks for an Authorization: header in 
incoming requests.

Other changes:

Added support for an HTTP302 exception, now used
by the auth/__init__.py when it needs to send a challenge.
The challenge will eventually include redirect information, 
but leaving that out for now. There is no code responding
to /challenge, yet.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5649 bb0f57cd-c710-0410-a8fb-e8bc1be0d679