tiddlyweb - first pass at user challenging and credentials extraction.
This is not the complete nor even pluggable version of how challenges
and credentials handling will eventually work in TiddlyWeb. This work is
to flesh out how the HTTP flow might work, and determine where in the
stack of middleware things need to happen.

The temporary status of this stuff is indicated with some comments and
things like cookie names being 'tiddlyweb_insecure_user'.

The basic flow goes like this:

Credentials Extraction: An incoming request is inspected, first for an
Authorization header in the 'Basic' style. If that's there, and the
username and password match (one another) tiddlyweb.usersign is set in
the environment. Then we look for a tiddlyweb_insecure_user cookie. If
that's there it's value is assigned to tiddlyweb.usersign. If we still
have nothing, tiddlyweb.usersign is set to GUEST.

When a bag is accessed for some operation, if the bag policy does not
allow that action on of two errors is raised. If the usersign is 'GUEST'
we throw a UserRequiredError. If the usersign is someone else, a
ForbiddenError is thrown.

These two errors are caught by a PermissionsExceptor middleware, which
turns the errors into something HTTP related.

ForbiddenError results in an HTTP 403 Forbidden and the client is left
to figure out how to move on from there (presumably they will reauth as
someone else).

If the request method is 'GET' a UserRequiredError redirects to
/challenge. If the method is something else a 403 Forbidden is sent. We
do things this way because flexible recovery from auth failure while
doing a POST, PUT or DELETE is tricky as we cannot guarantee that code
will maintain request state through a series of redirects.

Challenge: /challenge responds with a 401 and an HTML list of the
various challengers that are available. The only one right now is
simple_cookie. If the user or client follows that link, a form is
presented. Fill out the form with matching username and password,
submit, and tiddlyweb_insecure_user cookie is set in the headers, and
the client is redirect back to the location they were trying to get to
when the UserRequiredError was raised. Henceforth, new requests will
include the cookie so tiddlyweb.usersign will get set to the value of
the cookie.

The plan is that credentials extraction will be an ordered list of ways
to validate some arbitrary thing in the headers of an incoming request
that indicates a user. The list will be based on server configuration.
Each extractor either returns a usersign, or returns False. If the
entire list cascades as False, the usersign is set to GUEST.

Challengers will be an unordered list. All of them will be presented on
the /challenge page as a convenience to people access the tiddlyweb
server from their browsers. For people using TiddlyWikis that will need
to authenticate on the fly to the servers, it is assumed that some of
the time the TiddlyWiki will be self configured to choose a particular
challenger automagically and manipulate it directly (because using the
challengers interactively from with the TiddlyWiki will be tricky). A
bit more to work out there.

There is no necessary direct pairing between challengers and extractors.
Lots of different challengers could set tiddlyweb_insecure_user. It is
expected, for example, that we only want to hassle the openId server
once per session, so to avoid hassle, the end results of the interaction
with the openId server is a Set-Cookie: tiddlyweb_insecure_user (of
course using the secure version of such handling, to come)

The challenger and credentials modules will be loaded in much the same
way as the stores and serializations.

If you want to play with this you need to edit a bag's policy so that
read is limited to one or more users. Once that's done you can attempt
to access the tiddlers in that bag from a browser. You should get a
redirect to /challenge.

There will be a huge amount of work to get this cleaned up and up
to the standard of some of the earlier parts of TiddlyWeb. Much of the
code writing for this checkin is learning, speculation, and
experimentation. Which doesn't do much for its readability.

Comments desired. Thank you.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5653 bb0f57cd-c710-0410-a8fb-e8bc1be0d679