tiddlyweb - rudimentary support for Atom feeds of filtered bags and recipes, and single tiddlers, with tiddlers formatted to HTML.
This has been created (and demonstrated in atom_server.py) using an 
external, non-default, serialization, in atom/atom.py. A default
run of server.py or demo.py does not include the atom server. See
atom/README for a bit more information.

The feed generated is incomplete:

* it doesn't have bits of metadata that would be nice to have (such as
  dates and authors)
* it is only HTML, I can see how feeds of tiddlywikitext would also be
  useful. Not hard to do with a subclass.

But it shows up in Netnewswire and I can click on tiddlylinks and go
back to the server to the original content.

There's no support yet for service documents or any of that stuff so we
are some distance from APP. Wanted to start simple first. It should be
possible to do.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5722 bb0f57cd-c710-0410-a8fb-e8bc1be0d679