tiddlyweb - a bunch of tweaks to get test coverage back up to 96.9%.
Much of the rest of lack of coverage is artifacts from weirdness in the
coverage tool, so not going to stress it.

Only significant functional chnange is that tiddlyweb.web.util.server_base_url()
now does not include the trailing /. With it there, there are two many %s%s 
situations, which is just downright confusing. Better %s/%s.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5730 bb0f57cd-c710-0410-a8fb-e8bc1be0d679