tiddlyweb - Introduce the concept of a NoSerializationError and add list_tiddlers(bag) to the serialization interface.
This is a fairly large change.

It started out as a way of trying to introduce to simply horn
a <link> tag into HTML representations of tiddler lists when we
have support for atom in place. At first glance this should just
be a matter of overloading bag_as in the html Serialization.
That does in fact work, but it starts to expose little niggling
problems all over the place with the serialization of bags. Mostly
related to deciding whether or not we are going to do any for
a particular serialization: there was no reliable way to say
right now, under these circumstances I can't do that, while also
maintain the idea that we should be able to add and remove serializers
at will. So along came NoSerializationError. A serializer raises
that when it doesn't want to do what it has been asked to do
(for example, turn a bag into a wiki).

Originally bag_as in the serialization system would list 
tiddlers in whatever the format is supposed to be. Increasingly this
has seemed really smelly, because a bag is not a list of tiddlers,
it is a container for some tiddlers. So rather than continue to fight
it and screw around, I decided to get rid of the smell and stick
list_tiddlers into the serialization interface. This makes things
much more minty fresh _and_ allows (at our option) to put 
outgoing serializations of bags into the interface. That hadn't 
been possible in the past, which was going to be a bit weird
if we ever wanted to have client side bag editors and such things.

These changes touched the code all over the place, but in the
end I think things are more simple. The tests are passing but
there may be some lurking nasties.



git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5774 bb0f57cd-c710-0410-a8fb-e8bc1be0d679