tiddlyweb - Make it so stores get the environment when they are instantiated. This makes it possible to pass in configuration information, so that things like the text store know where to put content (rather than using a constant).
At the same time extend the server_store entry in the config
dict to be a two item list. The first item is the name of the
store, the second item is an optional dict with whatever you
want in it. That's later available in the environ.

And extract the config dict to its own file. Eventually this
will be something that reads an actual file. Shocking


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5776 bb0f57cd-c710-0410-a8fb-e8bc1be0d679