tiddlyweb - bare bones support for doing searching in the stores.
This is not hooked up to the URL map, yet.

A search(string) method is added to the StorageInterface.
A store may choose to support it or not. A store may use
the search string however it likes to hunt around in the store
for tiddlers that might match. It should return a list of 
lazy (no text or attributes beyond title and bag) tiddlers.
Calling code is responsible for doing permissions checking.

A tiddler_written(tiddler) method is added to the StorageInterface.
This is so stores which wish to maintain a complexi index, or
send an event to a search engine or otherwise notify something
that there is new content, they may do so, by overriding
tiddler_written.

The implementation in stores/text.py is intentionally naive
and hopelessly slow with a large collection of tiddlers. This
is to encourage other people to override the implementation
with their own.

Will hook this into the interface next.

git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5802 bb0f57cd-c710-0410-a8fb-e8bc1be0d679