tiddlyweb - connect search of the store to /search?q=<query> on the web. Support the usual serializations.
As expected, this was a bit of a mess. The usual mechanism for 
sending a list of tiddlers to an output to to make a bag and
then send that bag wherever, in a serialized form. Early bags
do de-duplication handling to make recipes perform in the desired
way. For search results we don't want to do de-deduping in the same
was as there could be tiddlers of the same name from several 
different bags.

So the structure of Bags and the handling of HTML and JSON output of
list of tiddlers had to be adjusted to code with this. There
may be bugs.

There is no search form. If you want to search, just go to
/search?q=<your query string>


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5803 bb0f57cd-c710-0410-a8fb-e8bc1be0d679