tiddlyweb - A rather rough and tumble OpenID challenger for tiddlyweb. This one very intentionally does not use any open id libraries so that the process of the protocol can be clearly exposed and the implementation can remain stupid. Many libraries are all "session this", "store that", "obscure the http" that and that was really getting me down.
So I did it raw, based on http://wiki.openid.net/Introduction

This implementation does not support delegation. It has only
been tested against version 1 servers (livejournal and aol
thus far). It currently uses openid.mode checkid_setup. It 
probably makes sense to use checkid_immediate. Comments?

There is _not_ an openid extractor. If the openid challenge
is valid, the Cookie used by the simple_cookie extractor is
Set to the value of openid.identity (with leading https?://
stripped). This means that servers that want to use openid
based access control need to put openids in bag policies, like this:

{"read": ["cdent.livejournal.com"], "create": [], "manage": ["NONE"], "write": [], "owner": null, "delete": []}

I suspect there are much better ways to implement OpenID.
This is the bare minimum to get someone rolling. The beauty
of the extensibility of the ChallengerInterface is that someone
else can write a better one and use it instead if they like.

Elsewhere in this commit are some cleanups to the reading
of POST or PUT input. We need to read(CONTENT_LENGTH) not
read(), otherwise we may not read at all.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@5851 bb0f57cd-c710-0410-a8fb-e8bc1be0d679