tiddlyweb - fixed and comments learned from setting up an authenticated wiki that runs as a CGI under apache (see index.cgi in the distribution).
The importer needs to effectively survive a tiddler that
has no contents. It also needs to deal with with tiddler
titles that the store can't save. Right now the importer
just warns and carries on. This is not sufficient, thus
a change in the TODO.

The challengers need to redirect to the right place after
they are done. When a server_prefix is used this can lead
to a mess as the server_prefix was getting added twice.
So add a server_host_url which is used by the previously
existing server_base_url. Now the challengers redirect
correctly.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@6258 bb0f57cd-c710-0410-a8fb-e8bc1be0d679