tiddlyweb - Implement deletion of bags in the StorageInterface and urls.map.
If you delete a bag all the tiddlers in it are _gone_.

This presents some interesting notions with regard to 
how search indexes are built and handled. For now we are
just ignoring those issues.

In order for a bag to be delete the delete constraint on 
the bag's policy must pass.

We need to handle the situation where a store doesn't support
deletings bags (or whatever else), so a soon commit will
implement some StoreMethodNotImplemented exceptions, or something
like that. The current hack for bag_delete is to recognize
when a pass has happened, which doesn't map to the style of
the rest of the system.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@6408 bb0f57cd-c710-0410-a8fb-e8bc1be0d679