tiddlyweb - make the wsgi middleware stack configurable from the tiddlywebconfig system. The first pass at this introduced a circular dependency, the resolution of which made for touching lots of files.
serve.default_app() is no longer needed, so it has been removed

What this change does is add two config items:

  'server_request_filters': [Query, StoreSet, UserExtract, Negotiate],
  'server_response_filters': [PermissionsExceptor, HTTPExceptor, EncodeUTF8, SimpleLog],

which provides an ordered list of the filters. The previous way the 
order was hard to discern.

By putting these things in config it is possible to do things like _not_
include SimpleLog, or to stick in middleware to do interesting stuff
not originally provided by TiddlyWeb. This is because the entries
in the lists (above) are proper class or function references. Code.
In your app stack. Do whatever you want!

addresses trac ticket: #737

this commit breaks the google app engine version. a fix for that will
be forthcoming



git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@6558 bb0f57cd-c710-0410-a8fb-e8bc1be0d679