tidldyweb - Finish up the implementation of support for user roles.
This required changing the contents of the tiddlyweb.usersign entry
in the wsgi environment from just the username, to a dict looking like:

    name: <name>
    roles: <list of roles>

Policies can now contains roles, indicated by a prefix of 'R:', so
for example, to give delete privs to cdent, jeremy and anyone with the
ADMIN role, make the delete constraints look like this:

   delete=['cdent','jeremy','R:ADMIN']

Credentials extractors now return a the dict (name, roles) described
above instead of a string, so the extractors have been updated accordingly.

Note, importantly, that the value in 'roles' can be anything the extractor
chooses: something from the User database, something it calculates, something
figured from the request. If it is not set, the policy will think the roles
list is empty.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@6580 bb0f57cd-c710-0410-a8fb-e8bc1be0d679