tiddlyweb - continued work on extended fields for tiddlers. This checking provides support in the JSON, HTML and wiki serializations. It is not as robustly tested as I would like. For the wiki and HTML serializations it's pretty much just a matter of atting attributes to the outgoing div, or in the case of the wiki div attending to those attributes which surprise us.
JSON just kind of works because all we are really doing is
change a python dict into a JSON dict and vice versa.

Note that none of these serializations try to do anything
with server.*. I've decided it the store that should worry
about that. As long as we never write server.* fields then
we won't send the wrong ones.

Expect bugs in the future!

Slope: confirm that these things work okay through the
web interface.

git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@6770 bb0f57cd-c710-0410-a8fb-e8bc1be0d679