tiddlyweb - provide access control for creating and writing bags (other permissions to come).
If the bag exists, we read its policy to see if the
current user is allowed to 'manage' the bag. If so,
they can write a new policy. If the bag does not
exist we check the bag_createp() function to see if
this user can create bags. At the moment this function
always returns true. Eventually it will do some
hoop hopping to raise Exceptions at the right times.
My expectation is that bag and recipe create policies
will be set in the config file and start out with something
like:

* anybody can
* any authenticated user can
* anybody with admin role can

git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@7061 bb0f57cd-c710-0410-a8fb-e8bc1be0d679