tiddlyweb - manage create permissions for bags and recipes.
Two new entires in config: bag_create_policy and
recipe_create_policy. These may each take one of 
three values:

'ADMIN': current user must have admin role
'ANY':   current user must be authenticated (not GUEST)
'':      anybody can create

A new function in the policy package handles 
the tests and raises appropriate exceptions.

Note that just because these polices are set it doesn't
mean that code running for a particular can't create a
bag, it just means that a user over the HTTP API can or
cannot. It's perfectly reasonable for the user to use a UI
that then administratively decides to create a bag for
the user.

git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@7067 bb0f57cd-c710-0410-a8fb-e8bc1be0d679