tiddlyweb - Add an experimental bit of editor functionality to TiddlyWeb. This is not a finished thing, just a demonstration of some ways in which you can do fun things with various parts of TiddlyWeb.
Any TiddlyWeb page which displays an HTML representation of a
tiddler (but not a tiddler revision) can optionally contain an
'Edit' link which points to an /edit url defined in urls.map.

edit takes query parameter of tiddler, bag and recipe. tiddler
and one of either bag or recipe is required.

Going to that link will populate a TiddlyWiki with the currently
viewed tiddler and create a few "virtual" tiddlers and stick them
into the created TiddlyWiki (see tiddlyweb/web/edit.py for details).

Finally, any tiddlers identified in
environ['tiddlyweb.config']['editor_tiddlers'] will also be pushed
into the tiddlywiki. Using notation like this in config:

    'editor_tiddlers': [
        ['bag0', 'TiddlyWebAdaptorPlugin'],
        ['bag0', 'TiddlyWebAutoSavePlugin'],
    ],

The presence of 'editor_tiddlers' in config also happens to be what
controls the optional 'Edit' link above. Unfortunately the Edit link
shows up all the time even if the current user doesn't have permission
to write to the current tiddler. It could check for that but I wanted
to keep things simple for now. 

The primary function of this experiment is to demonstrate programmatic
(rather than recipe) generation of dynamic TiddlyWikis for various 
purposes. If this proves useful or interesting we can improve the
way the functionality is provided.


git-svn-id: http://svn.tiddlywiki.org/Trunk/contributors/ChrisDent/experimental/TiddlyWeb@7156 bb0f57cd-c710-0410-a8fb-e8bc1be0d679