tiddlyweb - move urls.map and empty.html into the tiddlyweb package as what is known as package data. This allows us to distribute those files and not need to copy them to use them for a default installation.
Adjust config so that it knows where to look for those files. Adjust
tests for same reason.

Some cleanups to docs etc as well (lots more of that needed).

So that leaves up with being able to do the following steps:

* get tiddlyweb tarball
* unpack and cd in there
* sudo python setup.py install
* cd <some random dir>
* mkdir -p store/{recipes,bags,users}
* twanager bag foo
* twanager tiddler bar foo < /some/text
* twanager server 0.0.0.0 8080
* open http://0.0.0.0:8080/bags/foo/tiddlers.wiki

And get a TiddlyWiki.

This is assuming the dependency resolution in setup.py
goes okay, which ... works for me!


git-svn-id: http://svn.tiddlywiki.org/Trunk/association/serversides/tiddlyweb/core@7320 bb0f57cd-c710-0410-a8fb-e8bc1be0d679