tiddlyweb - that performance fix broke the feature that was slow. That happened because it didn't have accurate test coverage. Now it works, is still more performant than before and has better test coverage.
The problem was that the original change isntantiated a
fresh bag, not the one that was already floating around
in the process. We don't have access to that bag in
the current context, so we still need to load it, but we
can cache the results in an object local cache.

git-svn-id: http://svn.tiddlywiki.org/Trunk/association/serversides/tiddlyweb/core@8638 bb0f57cd-c710-0410-a8fb-e8bc1be0d679