tiddlyweb - adjust test to reflect that the openid challenger now sends a 302 when it used to send a 303.
This is more complicated than it first appears.

A POST followed by a 303 should result in a GET to the Location
headers. A POST followed by 302 should require the browser to
alert you to what it is doing. However most browsers don't actually
follow those rules. So we were using a 303 to be sure.

However, just now when using mod_wsgi under apache 2, it was swallowing
the Location header when sending a 303, resulting in openid challenges
just sitting there dumbly. Changing to 302 resolved the situation.

Sigh.

So therefore the test now needs to change because httplib2, being
a good little things, doesn't like a 302 after a POST.

Ooooph.

Or at least I think that's what's going on. Please correct me
if I'm wrong.

git-svn-id: http://svn.tiddlywiki.org/Trunk/association/serversides/tiddlyweb/core@8785 bb0f57cd-c710-0410-a8fb-e8bc1be0d679