tiddlyweb - include bag names in the instance_tiddlers information, this makes it possible to put different tiddlers in different bags or otherwise get things into various places using the instance and update commands on twanager.
This implementation is based on a great idea and inital path from
FND. This is different from his patch in that instead of using a
dict for the config entry it uses a list of two piece (bag, list
of tiddlers) tuples. This is because the config items that are
dicts can be extended whereas scalars, lists or tuples are replaced.

Also: add a comment to direct people to the docs in tiddlywebconfig.py

git-svn-id: http://svn.tiddlywiki.org/Trunk/association/serversides/tiddlyweb/core@9202 bb0f57cd-c710-0410-a8fb-e8bc1be0d679