tiddlyweb - Support HEAD requests as truncated GETs.
We now give a correct response when getting a HEAD: the request
is transformed into a GET and then the headers are sent without
the body. This is better than sending a 405 (Method Not Allowed).

As the cost of generating data is fairly high compared to high
bandwidth, there's no particular advantage to sending a HEAD
request to TiddlyWeb, but supporting it is useful.

This may be problematic on apache, but that will be revealed over
time. If it is problematic, the problems will only show up when
doing HEADs.