Remove the old filter code and fix everything the failing tests indicated should be fixed.
Now all the tests pass and we are using the new filter style throughout.

Documentation still needs to be updated, and one problem has been exposed:

It's hard, when filtering a bag, to just say "give me the three tiddlers
named foo, bar and baz". To do that in the current syntax you would need to
make three requests to the same bag.

Possible solutions:

/bags/zoom/tiddlers?select=title:foo,bar,baz

/bags/zoom/tiddlers?select=title:foo,title:bar,title:baz

But what, then, about ',' in titles (etc). And where do we draw the
line, is it only titles that can do this, or can all types of
selectors?

Comments please.