Adjust the tests so they can continue to work when the current server_store is not the text store.
This makes it possible to use the test suite for testing other stores.

Where possible the tests don't depend on file locations for their work.
Where they do, py.test.skip is used to skip them when the server_store
is not the tiddlyweb.stores.text store.

So, with these changes in place, if you make server_store different in
tiddlywebconfig.py before the tests are run, the tests will use the
other store.

Some of the tests still have file related things in them, so it will
take some work with non-file stores before this is fully settled out,
but is a good step forward: the simpletext store was tested against this.