Only instantiate once store reference per instance of the StoreSet WSGI middleware.
I was under the misguided notion that we were making a new instance
per request and that it went out of scope at the end. This is, however,
not the case. One instance gets created and is reused.

So what was happening was that a new Store object was being created
everytime __call__ is called, and never going out of scope. So what
we want to have happen instead is for Store to create one and for
us to use it over and over again. New processes or threads will
create other ones. This should have the effect of reusing things
like the sqlalchemy connection pool.