Start building the architecture for validating/sanitizing tiddler input on PUT.
The basic idea is that a bag policy now has an accept field, which means:
accept without sanitizing for people named in this field, otherwise sanitize.

As with the other contraints an empty list means accept for anybody, NONE means
accept for nobody (always sanitize), ANY means any authed user is not sanitized,
roles and usernames worked as expected.

When a tiddler is to be validated, it is passed, in sequence to any of the
validators defined in tiddlyweb.web.validator.TIDDLER_VALIDATORS. A validator
is function with the signature (tiddler, environ), environ is optional.

A validator either modifies the active tiddler, or raises InvalidTiddlerError.

If InvalidTiddlerError is raised, the user agent is sent an HTTP 409.