Fix issues with generators in filters, especially in limit.
Limit was trying to take a slice on a generator, which you
can't do, so first pass fix was to make the generator a list
and then take a slice. However the common case for using limit
is to take the first N of a list. In which case we may as well
just roll the existing generator, and stop iteration at the point
where we don't want the data.