Test and fix a bug with double decoding in the text store's search method.
This was discovered while testing the devtext store.

This search method is really crappy. Far too much implementation dependent
stuff in it that is duplicated with slight differences from elsewhere
in the same modules. This makes it impossible to inherit in any useful
fashion and makes it a bear to read.

I can feel an abstraction coming on. Someone stop me!