Make a bag just a list of tiddlers, without managing duplicates.
It turns out that duplicate handling is taken care of elsewhere
in the TiddlyWeb code and doing it within the Bag itself encumbers
the bag with a lot of expensive overheard that is not needed for
most cases.

Duplication handling is important when processing recipes, the
tiddlyweb.control module takes care of this.