Corrent ETag handling for new tiddlers.
If a new tiddler is PUT with an ETag we should refuse it,
because tiddlers that don't exist yet don't have Etags. What this
effectively does is make sure that clobber protection is also
operational for new tiddlers.

Also fix two rare problems when putting tiddlers:

If content-type or content-length are not provided on a PUT,
we should raise 400 not 500.