Correct handling of i18n data in submitted form data. Possible backwards incompatibility for some plugins.
tiddlyweb.query was not properly handling non-ascii data
in form submissions. An earlier fix for this was not even
remotely correct, so I bit the bullet, added some tests,
and fixed the problem for real.

The result is that data in tiddlyweb.query is now a unicode
string and form input is expected to be UTF-8 encoded. This is
much more inline with the unicode inside, UTF-8 outside principle
of TiddlyWeb.

This will likely break some plugins that do their own form handling.