Adjust HTML serialization to preserve navigation in the face of strict bag policies.
When a bag's manage policy is set to some users or roles, only
those users or roles can view the /bags/{bag_name} URL. Prior to
this change this mean that one could go to the /bags URL, click on
a link and get confronted with a challenger, despite the face that
/bags/{bag_name}/tiddlers would be readable if accessed directly.

It's better to be able to navigate. So the new system links to
/bags/{bag_name}/tiddlers from /bags and the tiddlers page links
to /bags/{bag_name}

The reason the manage policy limits who can view a bag is because
the usernames and roles in policies should not be exposed in secure
situations.