Second attempt at enabling the use of an index when doing a select filter.
The general idea here is that when a filter is created, the environ
is added to the data structure. Later when the filter is used, if
the incoming thing to be filtered is a bag where all the tiddlers are
from the same bag, then we have the option to try to use the index.
We then do a search against the index, where bag is bag name, and
we query from attribute:value.

This assume _many_ things:

* that we are using whoosher and can just import it
* the attribute we are matching is in the whoosher scheme
* that value we are matching is an exact term

These are probably not valid assumption, so some work is required.

To really get this to work properly the right thing to is probably:

* Make filters optionally a Class which is __call__able, thus
  sticking all this extra metadata on them.
* Formalize/reify the concept of an index system so it can be
  boostrapped more cleanly than this code currently does.