Tune up the indexed filtering system so it uses an indexer module.
An indexer module provides a callable called index_query which
takes environ and dictionary of key value pairs from which a
selection query is to be generated. The method returns a generator
of tiddlers.

index_query can do whatever it likes, but it is assumed it will
return tiddlers that match the query.