Continued modification to ETag handling to make up for browsers not being good at HTTP.
Many (most?) browsers do not attend to the Vary header properly,
so yesterday's addition of the vary header to deal with different
representations of the same resource being cached properly did not
take. To get around this I've adjusted the ETag generation so that
the mime type of the representation is added to the ETag.

When ETags are checked during write operations (DELETE and PUT) the
mime type is chopped off if it is there before match checks are made.
This means that if this change works, it will not require changes
in existing client code: the old type of ETag will continue to be
the correct form.

I'm only 50% that this is going to do the trick.