Fix a bug present when PUTting a tiddler to a recipe.
The correct behavior is that a tiddler will go in a bag with a filter on
it if the tiddler's attributes match the filter appropriately.

Turns out this never worked because the tiddler were not be deserialized
before the filter handling was being checked.

Now they are.

It's very surprising that this problem hasn't reared it's head before.
This suggests that the public use of TiddlyWeb thus far while fairly
broad has not been particular deep.

kudos to Jon Robson for finding the problem.