Discussions about using binary tiddlers as ways of managing static content pointed out that the code was not working as expected.
A binary tiddler was _always_ showing up as its binary form,
even if one asked for say a text or json rep. What should happen
is that the binary form should be provided when requesting the default
serializer or the a serializer that matches the type on the binary
tiddler. Otherwise the the serialization system handle. This means,
for example, that the JSON and Text serializers will set their text
fields to the base64 encoded value of tiddler.text.

While work on this it became clear that there was an inconsistency
with how the text store was managing binary content and how the
text serialization (which the text stores uses) was. The responsiblity
was in the wrong place, so now it has been moved.