Integrate select=field:<fieldname> type select filters.
This is based on work done with RakugoJon to create filters which
say: "give me all the tiddlers that have the named field...so that
I can then sort on them".

Fields have become increasingly important in the way people are using
tiddlers, so making this core seems to make sense.