Make some adjustments to binary tiddler handling.
A request for an extension-less tiddler, using an Accept header
of application/json was not returning json. Was returning whatever
the mime-type of the tiddler.

This may break handling of some rare external serializers, however
that specific case doesn't appear to have test coverage so we are
using the real world to re-expose the bug (if it still exists) so
we can get proper coverage on it.