Fix a painfully subtle bug with config handling.
When config is imported from tiddlyweb.config the first time,
if tiddlywebconfig.py is present, the config in there is merged
to the main config. Later if merge_config is called again with
yet more data, tiddlywebconfig.py's data is supposed to be
read in again and continue to take precedence.

This was not working for keys which did not already existing
in tiddlyweb.config and for which the value (of the new
config item) was a dict.

The diff will illuminate the fix. Basically a reference rather
than a value was getting passed around when it shouldn't.

The commit of tiddlywebconfig.py is required to properly test
the situation in something resembling a real world situation.